kubepkg: Move build definition templating logic to template.go
(No logic changes here, just code organization.)
